refactor(schematic): parse json
- refactor and extend the CLI package json method to accept a path
